backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI/ML
anthropic==0.68.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
PyPDF2==3.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2